ORM/Achha/requirement.txt
asgiref==3.5.2
Django==4.0.6
django-session-timeout==0.1.0
djangorestframework==3.13.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
